requirements.txt
feedparser==6.0.10
PyYAML>=6.0
requests>=2.31.0